refactor: restructure pages folders
